WIP: draft for delegation events
