+++ contract-bindings-ethers/src/lib.rs
@@ -13,3 +13,4 @@ pub mod permissioned_stake_table;
 pub mod plonk_verifier;
 pub mod plonk_verifier_2;
 pub mod shared_types;
+pub mod stake_table;
