+++ contract-bindings-ethers/src/shared_types.rs
@@ -15,6 +15,42 @@ pub struct G1Point {
     pub x: ::ethers::core::types::U256,
     pub y: ::ethers::core::types::U256,
 }
+///`G2Point(uint256,uint256,uint256,uint256)`
+#[derive(
+    Clone,
+    ::ethers::contract::EthAbiType,
+    ::ethers::contract::EthAbiCodec,
+    serde::Serialize,
+    serde::Deserialize,
+    Default,
+    Debug,
+    PartialEq,
+    Eq,
+    Hash,
+)]
+pub struct G2Point {
+    pub x_0: ::ethers::core::types::U256,
+    pub x_1: ::ethers::core::types::U256,
+    pub y_0: ::ethers::core::types::U256,
+    pub y_1: ::ethers::core::types::U256,
+}
+///`EdOnBN254Point(uint256,uint256)`
+#[derive(
+    Clone,
+    ::ethers::contract::EthAbiType,
+    ::ethers::contract::EthAbiCodec,
+    serde::Serialize,
+    serde::Deserialize,
+    Default,
+    Debug,
+    PartialEq,
+    Eq,
+    Hash,
+)]
+pub struct EdOnBN254Point {
+    pub x: ::ethers::core::types::U256,
+    pub y: ::ethers::core::types::U256,
+}
 ///`PlonkProof((uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),(uint256,uint256),uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)`
 #[derive(
     Clone,
